+++ example/ecommerce/internal/types/customers/types.go
@@ -33,16 +33,16 @@ type CreateCustomerReq struct {
 }
 
 type Customer struct {
-	ID             shared.NonEmptyID     `json:"id"`
-	Email          shared.Email          `json:"email"`
-	Name           string                `json:"name"`
-	PrimaryAddress Address               `json:"primaryAddress"`
-	BillingAddress *Address              `json:"billingAddress,omitempty"`
-	Contact        Contact               `json:"contact"`
-	Avatar         *multipart.FileHeader `json:"avatar,omitempty"`
-	LoyaltyPoints  *int                  `json:"loyaltyPoints"`
-	Bio            *string               `json:"bio,omitempty"`
-	CreatedAt      string                `json:"createdAt"`
+	ID             shared.NonEmptyID `json:"id"`
+	Email          shared.Email      `json:"email"`
+	Name           string            `json:"name"`
+	PrimaryAddress Address           `json:"primaryAddress"`
+	BillingAddress *Address          `json:"billingAddress,omitempty"`
+	Contact        Contact           `json:"contact"`
+	AvatarURL      *string           `json:"avatarUrl,omitempty"`
+	LoyaltyPoints  *int              `json:"loyaltyPoints"`
+	Bio            *string           `json:"bio,omitempty"`
+	CreatedAt      string            `json:"createdAt"`
 }
 
 type CustomerRef struct {